pushed view stuff

+++ src/model/ModelEvent.java
@@ -4,10 +4,10 @@ public class ModelEvent {
     private double user_balance;
     private AccountModel user;
     private double amount_transferred;
-    private int operations_completed;
+    private double operations_completed;
 
     
-    public ModelEvent(AccountModel acc, double a, int c) {
+    public ModelEvent(AccountModel acc, double a, double c) {
         this.user = acc;
         this.user_balance = acc.getBalance();
         this.amount_transferred = a;
@@ -20,6 +20,6 @@ public class ModelEvent {
     public void setAccount(AccountModel user) { this.user = user; }
     public double getAmount_transferred() { return amount_transferred; }
     public void setAmount_transferred(double amount_transferred) { this.amount_transferred = amount_transferred; }
-    public int getOperations_completed() { return operations_completed; }
+    public double getOperations_completed() { return operations_completed; }
     public void setOperations_completed(int operations_completed) { this.operations_completed = operations_completed; }
 }
